requirements.txt
matplotlib==3.7.5
numpy==1.24.3
numpy==1.24.1
opencv_python==4.9.0.80
Pillow==10.3.0
timm==0.5.4
tools==0.1.9
torch==2.2.2+cu118
torchvision==0.17.2+cu118
tqdm==4.66.2
